+++ src/api/test.rs
@@ -2145,6 +2145,7 @@ fn log_q_exp_overflow() {
     variance_boost_strength: None,
     variance_boost_deep: None,
     quant_rounding_bias: None,
+    ssim_rdmult_strength: None,
     max_pixel_count: 67_108_864,
     speed_settings: SpeedSettings {
       multiref: false,
@@ -2234,6 +2235,7 @@ fn guess_frame_subtypes_assert() {
     variance_boost_strength: None,
     variance_boost_deep: None,
     quant_rounding_bias: None,
+    ssim_rdmult_strength: None,
     max_pixel_count: 67_108_864,
     speed_settings: SpeedSettings {
       multiref: false,